Chapter 17

+++ pyt2_allfiles/python/exercises/ch17/movies/ui.py
@@ -3,26 +3,31 @@
 import db
 from objects import Movie
 
+
 def display_title():
     print("The Movie List program")
-    print()    
+    print()
     display_menu()
+
 
 def display_menu():
     print("COMMAND MENU")
     print("cat  - View movies by category")
     print("year - View movies by year")
+    print("min - View movies by under maximum minutes")
     print("add  - Add a movie")
     print("del  - Delete a movie")
     print("exit - Exit program")
-    print()    
+    print()
+
 
 def display_categories():
     print("CATEGORIES")
-    categories = db.get_categories()    
+    categories = db.get_categories()
     for category in categories:
         print(str(category.id) + ". " + category.name)
     print()
+
 
 def display_movies(movies, title_term):
     print("MOVIES - " + title_term)
@@ -33,7 +38,8 @@ def display_movies(movies, title_term):
         print(line_format.format(str(movie.id), movie.name,
                                  str(movie.year), str(movie.minutes),
                                  movie.category.name))
-    print()    
+    print()
+
 
 def display_movies_by_category():
     category_id = int(input("Category ID: "))
@@ -44,43 +50,58 @@ def display_movies_by_category():
         print()
         movies = db.get_movies_by_category(category_id)
         display_movies(movies, category.name.upper())
-    
+
+
 def display_movies_by_year():
     year = int(input("Year: "))
     print()
     movies = db.get_movies_by_year(year)
     display_movies(movies, str(year))
 
+
+def display_movies_by_minutes():
+    minutes = int(input("Maximum number of minutes: "))
+    print()
+    movies = db.get_movies_by_minutes(minutes)
+    display_movies(movies, str(minutes))
+
+
 def add_movie():
-    name        = input("Name: ")
-    year        = int(input("Year: "))
-    minutes     = int(input("Minutes: "))
+    name = input("Name: ")
+    year = int(input("Year: "))
+    minutes = int(input("Minutes: "))
     category_id = int(input("Category ID: "))
-    
+
     category = db.get_category(category_id)
     if category == None:
         print("There is no category with that ID. Movie NOT added.\n")
-    else:        
+    else:
         movie = Movie(name=name, year=year, minutes=minutes,
                       category=category)
-        db.add_movie(movie)    
+        db.add_movie(movie)
         print(name + " was added to database.\n")
+
 
 def delete_movie():
     movie_id = int(input("Movie ID: "))
-    db.delete_movie(movie_id)
-    print("Movie ID " + str(movie_id) + " was deleted from database.\n")
-        
+    movie = db.get_movie(movie_id)
+    if input(f"Are you sure you want to delete '{movie.name}'? (y/n)").lower() == "y":
+        db.delete_movie(movie_id)
+    print(f"'{movie.name}' was deleted from database.\n")
+
+
 def main():
     db.connect()
     display_title()
     display_categories()
-    while True:        
+    while True:
         command = input("Command: ")
         if command == "cat":
             display_movies_by_category()
         elif command == "year":
             display_movies_by_year()
+        elif command == "min":
+            display_movies_by_minutes()
         elif command == "add":
             add_movie()
         elif command == "del":
@@ -92,6 +113,7 @@ def main():
             display_menu()
     db.close()
     print("Bye!")
+
 
 if __name__ == "__main__":
     main()
